requirements.txt
#requirements.txt
Django==4.2.7
Pillow>=12.0.0
django-solo==2.1.0
django-admin-sortable2==2.1.4
django-cleanup==8.1.0
requests==2.32.3
python-dotenv==1.0.0
psycopg2-binary==2.9.11
django-simple-history==3.4.0

gunicorn==21.2.0          # WSGI сервер
whitenoise==6.6.0         # Обслуживание статики
django-redis==5.4.0       # Кеширование (опционально)
django-storages==1.14.3   # Для S3/облачного хранилища
boto3==1.34.84           # Для AWS S3